requirements.txt
google-api-python-client==1.8.0
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.1
psycopg2==2.8.5